Change 'unsigned int' to size_t in load() methods
The save methods were writing out <object>.size() which is of type
size_t but the load functions were reading in unsigned ints.  size_t is
64 bits while unsigned int is 32 so it was reading the wrong values.

+++ src/strategic/FMap.cpp
@@ -336,7 +336,7 @@ const int FMap::save(std::ostream &os) const{
 int FMap::load(std::istream &is){
 //	std::cerr << "Entering FMap::load" << std::endl;
 	read(is,m_maxCoord);
-	unsigned int sysCount,jumpCount;
+	size_t sysCount,jumpCount;
 	read(is,sysCount);
 	for (unsigned int i = 0; i < sysCount; i++){
 		FSystem *s = new FSystem;
